densplot {coda}	R Documentation

Probability density function estimate from MCMC output

Description

Displays a plot of the density estimate for each variable in x, calculated by the density function.

Usage

densplot(x, show.obs = TRUE, bwf, 
                ylim, xlab, ylab = "", type="l", main, ...)

Arguments

x
An mcmc or mcmc.list object

show.obs
Show observations along the x-axis

bwf
Function for calculating the bandwidth. If omitted, the bandwidth is calculate by 1.06 times the minimum of the standard deviation and the interquartile range divided by 1.34 times the sample size to the negative one fifth power

ylim
Limits on y axis. See plot.window

xlab
X-axis label. By default this will show the sample size and the bandwidth used for smoothing. See plot

ylab
Y-axis label. By default, this is blank. See plot

type
Plot type. See plot

main
An overall title for the plot. See title

...
Further graphical parameters

Note

You can call this function directly, but it is more usually called by the plot.mcmc function.

If a variable is bounded below at 0, or bounded in the interval [0,1], then the data are reflected at the boundary before being passed to the density() function. This allows correct estimation of a non-zero density at the boundary.

See Also

density, plot.mcmc.

[Package coda version 0.18-1 Index]